+++ Changelog.md
@@ -1,3 +1,12 @@
+1.1.0
+-----
+* Changed LICENSE.md
+* Changed ReadMe.md
+* Files generated does not have anymore "\#encoding: utf-8" on its first line
+* New options added: --full, --rspec-only, --cucumber-only, --minitest-only, --clean
+* Rakefile content is generated according to the options passed
+* Improved feedback from cocot
+
 1.0.1
 -----
 * Changed LICENSE.md